pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cugo-rs485-motor-control"
version = "0.1.0"
description = "RS485/Modbus RTU motor control tools for CuGoV4 on Raspberry Pi 4"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pyserial>=3.5",
]

[tool.setuptools.packages.find]
where = ["scripts"]
